+++ ktrain/utils.py
@@ -493,6 +493,18 @@ def list2chunks(a, n):
     return (a[i * k + min(i, m):(i + 1) * k + min(i + 1, m)] for i in range(n))
 
 
+def get_hf_model_name(model_id):
+    parts = model_id.split('/')
+    if len(parts) == 1:
+        model_id = parts[0]
+    else:
+        model_id = '/'.join(parts[1:])
+    if model_id.startswith('xlm-roberta'): 
+        model_name = 'xlm-roberta'
+    else:
+        model_name = model_id.split('-')[0]
+    return model_name
+
 
 class YTransform:
     def __init__(self, class_names=[], label_encoder=None):
@@ -634,6 +646,16 @@ class YTransformDataFrame(YTransform):
         return self.apply(df, train=True)
     def apply_test(self, df):
         return self.apply(df, train=False)
+
+
+
+
+
+
+
+
+
+